Removed 4 unnecessary stubbings in ChannelModeCommandTest.java
